Move permission helper method closer to where it's used
The `BatchInvitationPermissionsHelper#formatted_permission_name` method
is only used in `UsersHelper#items_for_permission_checkboxes` and so it
makes more sense to define it in `UsersHelper` closer to where it's
being used. We _could_ also make it private, but I decided it was worth
keeping the tests around it and so left it as public.