Add Doorkeeper::AccessToken.expires_after scope
Trello: https://trello.com/c/XfBBRL5M

My motivation for adding this is that I want to use it in
`SSOPushCredential.credentials`.

I've had to re-open the class from the doorkeeper gem, but we already
have precendent for doing that for `Doorkeeper::Application`.

I've based the implementation on
`Doorkeeper::Orm::ActiveRecord::Mixins::AccessToken::ClassMethods#not_expired`
- see this code [1]. Re-using the `#expiration_time_sql` method should
make the implementation robust against a change of database (e.g.
MySQL -> PostgreSQL).

Note that I've had to wrap the call to `#expiration_time_sql` in a call
to `ActiveRecord::Sanitization::ClassMethods#sanitize_sql` [2] in order
to avoid Brakeman failing with a SQL Injection warning.

[1]: https://github.com/doorkeeper-gem/doorkeeper/blob/986115cc228ff30dc1ead0f4101195448994f5d4/lib/doorkeeper/orm/active_record/mixins/access_token.rb#L47-L66
[2]: https://api.rubyonrails.org/v7.0.8/classes/ActiveRecord/Sanitization/ClassMethods.html#method-i-sanitize_sql