Don't exclude SSO Push tokens from Prometheus alert
In this commit [1], we excluded these tokens, because a new one was
created automatically on-demand if the old one had expired.

However, now that we're doing this 4 weeks before the token is due to
expire, it seems sensible to reinstate the alerting for tokens
associated with the SSO Push API user so that if the automatic creation
fails for some reason, we still get alerted.

It also has the benefit of making the code a bit simpler and easier to
understand.

[1]: c79c39a8c3dd20c63e6f0db9a046507c2fd0f810