Improve Organisation select on /users/invitation/new
Rosa got in touch to say that this Organisation select box is not very
usable, particularly in integration. The organisations are sorted in a
seemingly random order and they include closed organisations that we
never need to assign to new users.

This commit fixes both of those issues.

The previous tests stubbed `policy_scope` which is a Pundit-provided
convenience for `Pundit.policy_scope` automatically passing in
`current_user` as the first argument. I wanted to actually call the
scope, so I could test that the `not_closed` and `order` scopes I've
added were working correctly. To get `policy_scope` to work in the
isolated tests I've explicitly included `Pundit::Authorization` in
`UsersHelper` and stubbed `current_user` instead.