Use option-select vs checkboxes for granting permissions
Trello: https://trello.com/c/SfIc6Q2q

The main motivation for these changes is the Licensing app which has a
large number of permissions. The recent changes [1,2] which moved the UI
for granting permissions to the GOV.UK Design System have made managing
the permissions for the Licensing unmanageable.

This replaces the accordion component [3] where each item of the
accordion was a checkboxes component [4] (i.e. a set of checkboxes) with
a list of option-select components [5] (each of which has a set of
checkboxes within it). We've made this change to *both* the new
invitation ("Create new user") page and the new batch invitation
permissions page.

The option-select component gives us something similar to the show/hide
toggle which was previously provided by the accordion which is why it no
longer makes sense to use the accordion. Also it makes the permissions
section more compact which is no bad thing given how long it is! Most
importantly the option-select component gives us the option to enable a
search filter for the permissions for a given app. I plan to make use of
this in a subsequent commit. This will address the primary motivation
which is the unmanagability of the Licensing app permissions.

Note that I've set the `key` attribute of the option-select components
to "not-used" for now, because it's a required attribute but the name of
the checkboxes is being derived from each individual option hash. I plan
to address this in a subsequent commit.

The use of the option-select component isn't really what it was intended
for in that it doesn't *control* a list of things. However, I don't
think it's too far away and it gives us a much better UI.

[1]: https://github.com/alphagov/signon/pull/2361
[2]: https://github.com/alphagov/signon/pull/2412
[3]: https://components.publishing.service.gov.uk/component-guide/accordion
[4]: https://components.publishing.service.gov.uk/component-guide/checkboxes
[5]: https://components.publishing.service.gov.uk/component-guide/option_select