Only open option-selects when app has checked permissions
Set each option-select component to open on load only if any of its
permissions are selected. At the moment this is only relevant to the new
invitation page when validation errors have occurred and the form has
been re-rendered. The new batch invitation permissions page currently
never has any validation errors, so we can just hard-code the value of
`closed_on_load` to `true`. This behaviour is very similar to what we
did on the users index page.

Since this behaviour is implemented in JS, it's not possible to test it
in the controller test and it doesn't seem important enough to add an
integration test for it.