Put checked permissions at top on new invitations page
This change ensures any checked permission checkboxes are displayed at
the top of the list of checkboxes for a given application option-select.

At the moment this is only relevant to the new invitation page when
validation errors have occurred and the form has been re-rendered.
Currently the new batch invitation permissions page can never have any
validation errors.

This behaviour is very similar to what we did on the users index page.