Assert retired/non_retired apps are in relevant tab
Applications are listed in two tabs on the apps index page. This checks
that non-retired apps are displayed in the "active" tab and retired apps
are displayed in the "retired" tab. I'm planning to make a change in
this area and I wanted to ensure there was some test coverage in place.