Fix bug in UserUpdate to prevent loss of permissions
Trello: https://trello.com/c/boKPl89S

We use the UserUpdate class in `Account::RoleOrganisationsController` to
update the user's role and organisation. We're not passing
`supported_permission_ids` in the `params` parameter to
`UserUpdate#initialize` because we're not interested in changing a
user's permissions in these operations. Unfortunately it turns out that
`UserUpdate` expects to have the user's existing permissions passed in
if you want them to remain unchanged. Otherwise
`UserUpdate#filtered_user_params` will reset/clear all permissions for
that user.

I've chosen to fix this in `UserUpdate#filtered_user_params` because it
seems reasonable to me that leaving `supported_permission_ids` out of
`params` should be interpreted as meaning that supported permissions
shouldn't be changed.

Note that this bug would've caused superadmins to lose their permissions
if they changed their role or organisation, and admins to lose their
permissions if they changed their organisation.

In fixing this problem I discovered that `ManagePermissionsController`
was relying on the previous behaviour to allow users to remove all their
permissions. I've solved this in the same way that it's solved in
`UsersController#update` by copying the `allow_no_application_access`
method from that controller.