Record EventLog when organisation is changed
We appear to record an event when anything else is changed by
`UserUpdate` so it seems reasonable that we should also record an event
when the organisation is changed.

Superadmins and Admins can change users' organisations from or to "None"
so I've extracted the `Organisation::NONE` constant to avoid some
duplication between the template and the tests.