Introduce PreselectedPermission model
This model holds the permissions that we want to be pre-selected when
a new User is invited via /users/invitation/new. This list of
permissions has been given to us by the team that frequently creates
new users. Not all users will need to have these permissions, so we
don't want them to be automatically created on User creation, but
they are added frequently enough that having them preselected in the
new user invitation form will save some time and effort.

As these applications/permissions are stored in the database we can't
guarentee they exist at run time, and we wouldn't want to create them
if they have been removed for some reason so this class checks for
their presence.

We'll use this model to help build a new user in a subsequent commit.