Add tests for editing a retired application
I'm about to add a default scope for `Doorkeeper::Application` which
will mean that retired applications are not accessible from the UI by
default. I think this is one of the places in the app where we *do*
still want to have access to retired apps, so I'm adding this test
coverage in advance of adding the default scope to make sure I don't
break anything.