Add test for granting permissions to access retired app
I'm planning to add a default scope to `Doorkeeper::Application` which
will make retired apps unavailable by default. I suspect that will
change the behaviour captured in this new test and so having the test in
place first will make it clearer how the behaviour has changed in the
subsequent commit(s).