Preselect default permissions on /users/invitation/new
This commit ensures that the default permissions defined by the
SupportedPermission.default scope are checked when invititing a new
user. These permissions are currently automatically added when
inviting a new user but that mechanism is not very clear. By
preselecting them we give the inviting user the option to configure
them or add new ones via the application permissions UI. In a
subsequent commit I'll remove the code that adds these automatically
so everything is consistent.

I've inlined the implementation of Devise::InvitationsController#new
into InvitationsController#new so that I can set `self.resource` with
`User.with_preselected_permissions` instead of `User.new`.