Preselect some permissions on /users/invitation/new
This commit ensures that the preselected permissions defined in
PreselectedPermission are checked when invititing a new user. These
permissions are commonly (but not always) added when inviting a new
user so by having them preselected we save the inviting user some
time.

I've inlined the implementation of Devise::InvitationsController#new
into InvitationsController#new so that I can set `self.resource` with
`User.with_preselected_permissions` instead of `User.new`.