Add not_retired to Doorkeeper::Application default scope
Retiring an app is effectively a "soft delete" and so retired apps
should only appear in a very few pages in the UI, i.e. pages where the
user explicitly wants to view retired apps and pages where the user want
to edit a retired app, e.g. to un-retired it.

My motivation for doing this is from this Trello card [1] where we want
to reduce the number of apps displayed on various pages to only include
the useful ones. That card is actually about hiding API-only apps, but
it makes sense to me to hide retired apps from most pages as a first
step.

Changing the default scope like this broke the
`User#grant_application_permissions` method, because
`User#grant_permission` ended up trying to create a
`UserApplicationPermission`, but failed because the application was
`nil`. I've chosen to "fix" this method by changing its behaviour so it
no longer grants permissions for a *retired* app. I did consider raising
an exception if the app was retired, but in the end I decided that was a
bit drastic.

I've made use of `Doorkeeper::Application.unscoped` in the
`DoorkeeperApplicationsController` and its view templates to ensure that
the user can still view and edit retired apps.

It's possible that this commit will break some untested behaviour.
However, we should find out about that relatively quickly and I think
the benefits  outweigh the risks.

Note that I've had to tweak the `after :create` hook in the
`application` factory to avoid a validation error in a test. I _think_
this is because the application in the test is retired and when the hook
is executed `app` doesn't have an `id` which somehow results in the
relatively recently added presence validation for
`Doorkeeper::Application#application` failing. I only actually needed to
change the code in one of the loops within the hook to fix the test, but
it makes sense to me that I should change all of them at the same time.
I haven't investigated this very thoroughly, because (a) I think the new
version of the code is more idiomatic Rails; and (b) I really dislike
these FactoryBot hooks and I hope to get rid of them sometime soon!

[1]: https://trello.com/c/kvmb5OHO