Select default permissions on batch invite permissions form
This commit ensures that the default permissions defined by the
SupportedPermission.default scope are checked when adding a batch of
new users. These permissions are currently automatically added when
inviting batches of users but that mechanism is not very clear. By
preselecting them we give the inviting user the option to configure
them or add new ones via the application permissions UI. In a
subsequent commit I'll remove the code that adds these automatically
so everything is consistent.