Exclude retired apps from User#supported_permissions
We don't want to display a user's permissions for retired apps on any
page and so to ensure that I've introduced this `joins(:application)`
scope to the `User#supported_permissions` `has_many` association so that
it picks up the default scope on `Doorkeeper::Application`, i.e.
`not_retired`.

I don't think the `joins` scope itself will change which
`SupportedPermission` records are returned, because
`supported_permissions.application_id` has a `NOT NULL` and a foreign
key constraint and `SupportedPermission#application` has a `presence`
validation on it. The `joins` scope just serves to include the default
scope on `Doorkeeper::Application` in the query.