Add FK constraint to user_application_permissions.application_id
Fortunately this column already had a NOT NULL constraint on it and
`UserApplicationPermission#application` has a presence validation on it.

However, even so there were 20 records in the
`user_application_permissions` table in integration (but 0 records in
staging or production) where `application_id` did not reference the ID
of a record in the oauth_applications table!

I found these records using the following SQL query:

    SELECT user_application_permissions.id, user_application_permissions.application_id, oauth_applications.id
      FROM user_application_permissions
      LEFT OUTER JOIN oauth_applications ON user_application_permissions.application_id = oauth_applications.id
      WHERE oauth_applications.id IS NULL;

I've taken the same approach as in this commit [1] to fixup the data in
integration, i.e. I've used a Rails console to create a dummy
application for the other records to reference as follows:

    application_ids = ApplicationRecord.connection.query(%(
      SELECT user_application_permissions.application_id
        FROM user_application_permissions
        LEFT OUTER JOIN oauth_applications ON user_application_permissions.application_id = oauth_applications.id
        WHERE oauth_applications.id IS NULL;
      )
    ).flatten.uniq

    class ::Doorkeeper::Application < ActiveRecord::Base
      def create_signin_supported_permission; end
    end

    application_ids.each do |application_id|
      Doorkeeper::Application.create!(
        id: application_id,
        name: "Unknown #{application_id}",
        redirect_uri: "https://example.com/unknown-#{application_id}",
        description: "Added manually to give orphan user_application_permissions a parent application",
      )
    end

Running the above in my local copy of the integration environment fixed
the data so I could successfully run the migration to add the foreign
key constraint.

Note that the Rails migration has also automatically added an index on
`user_application_permissions.application_id` which seems sensible.

[1]: https://github.com/alphagov/signon/commit/747c8ed64fee1f80bf1f5afc4775cef36a6e08f7