Add presence validation for SupportedPermission#application
Now that we have a NOT NULL constrain on the
`supported_permissions.application_id` column, it seems sensible to add
this - it would be more user-friendly to get a validation error than a
MySQL exception if the application was not set.

Luckily the `supported_permission` factory already provided an
application by default, so this didn't break any tests.