Replace Application#show_on_dashboard -> #api_only
Trello: https://trello.com/c/kvmb5OHO

The `api_only` is effectively the inverse of `show_on_dashboard` and so
the migration creates a new column for `api_only` and populates it with
the inverse of the values in the old column, `show_on_dashboard`.

I think `api_only` better represents what this flag was intended to be
used for when it was introduced in this PR [1], although it's far from
clear what the actual intention was! Either way, it definitely better
represents what *I* plan to use it for!

. I've added some "hint" text alongside the checkbox to explain
a bit more what it means:

> An API-only application is one that doesn't have a user interface and
> is only used by API users. It will not appear on the dashboard.

[1]: https://github.com/alphagov/signon/pull/843