Add tests for GOV.UK admins to Account::ApplicationsControllerTest
While I think there is effectively coverage of this behaviour in the
`AccountApplicationsTest`, I found it confusing and potentially
problematic that the controller test only included tests with negative
assertions. I think it's important to always have tests that have the
equivalent positive assertions nearby - otherwise the negative
assertions might just be passing by accident.  Also in this case, I
think it makes it more obvious that the behaviour differs depending on
whether you are a GOV.UK admin or a Publishing Manager.