Do not display API-only apps on account applications page
Trello: https://trello.com/c/kvmb5OHO

Non-API users should not need to manage permissions for API-only apps
and so we can reduce the noise on this page by excluding API-only apps
as we already do for retired apps.

Note that we'll still be able to manage API user permissions on the API
users edit page.