Do not display API-only apps on a bunch of pages
Trello: https://trello.com/c/kvmb5OHO

Non-API users should not need to manage permissions for API-only apps
and so we can reduce the noise on these pages by excluding API-only apps
as we already do for retired apps. This change definitely affects the
following pages:
* `/users/invitation/new`
* `/batch_invitations/:batch_invitation_id/permissions/new`
* `/account/manage_permissions`

Changing `UserPermissionManageableApplicationPolicy#applications` is a
bit scary, because it's effectively used in quite a few places. In
particular this policy is used from
`SupportedPermissionPolicy::Scope#resolve` [1], but that only seems to
be used in relation to filtering allowed permission-related params [2]
for editing a non-API user [3], i.e. it won't affect editing permissions
for API users [4] which is what I was worried about. Anyway I'm fairly
confident we've got sufficient test coverage in place to catch any
problems that may have been introduced.

Note that we'll still be able to manage API user permissions on the API
user edit page and we have a test in place in `ApiUsersControllerTest`
that it remains the case. The code in this area is quite hard to follow,
but the reason that the change to
`UserPermissionManageableApplicationPolicy#applications` doesn't prevent
API-only apps from appearing on the API user page is that there is an
`if user.api_user?` condition in
`UserPermissionsControllerMethods#visible_applications` [5] which uses
some logic [6] not involving
`UserPermissionManageableApplicationPolicy`.

[1]: https://github.com/alphagov/signon/blob/3c3f5506dddf8e33dd5760d022c4d177f050398c/app/policies/supported_permission_policy.rb#L7
[2]: https://github.com/alphagov/signon/blob/3c3f5506dddf8e33dd5760d022c4d177f050398c/lib/supported_permission_parameter_filter.rb#L26
[3]: https://github.com/alphagov/signon/blob/3c3f5506dddf8e33dd5760d022c4d177f050398c/app/services/user_update.rb#L35
[4]: https://github.com/alphagov/signon/blob/3c3f5506dddf8e33dd5760d022c4d177f050398c/app/controllers/api_users_controller.rb#L39
[5]: https://github.com/alphagov/signon/blob/3c3f5506dddf8e33dd5760d022c4d177f050398c/lib/user_permissions_controller_methods.rb#L5
[6]: https://github.com/alphagov/signon/blob/3c3f5506dddf8e33dd5760d022c4d177f050398c/lib/user_permissions_controller_methods.rb#L6-L12