Ensure retired apps do not appear on user edit page
I'm about to make changes in this area and so I want to make sure there
is some test coverage beforehand.

Note that this is only relevant when the current user is a publishing
manager, because that's the only scenario in which the "App permission
for this user" table is displayed.

Also I have to use the `user_application_permission` factory to give the
edited user permission to the retired app, because using the `user`
factory with the `with_permissions` option doesn't work for retired
apps.