Add test for permissions fields on API user edit page
This may be covered in an integration test, but I'm about to make
changes which could affect this behaviour and so I want some low-level
coverage in place beforehand.

I find it a bit confusing that there is a `if
@api_user.authorisations.present?` condition around the
`app/views/shared/_user_permissions.html.erb` partial in
`app/views/api_users/edit.html.erb` given that the partial is displaying
permissions and not authorisations.

I know that _normally_ an API user is given signin permission to an app
when they are given an access token for that app [1], but I still think
it would be less confusing if this condition was `if
@api_user.supported_permissions.any?` or something along those lines.

Anyway, even though it wasn't strictly necessary to get the test to pass
I've given the API user signin permission to the app as well as giving
them an access token.

[1]: https://github.com/alphagov/signon/blob/3c3f5506dddf8e33dd5760d022c4d177f050398c/app/controllers/authorisations_controller.rb#L20