Ensure managing permissions of retired apps not allowed
I'm planning to make changes that might inadvertently change this
behaviour and so I want to make sure there is some test coverage in
place beforehand.

Note that the `UserPermissionManageableApplicationPolicy::Scope` does
not affect the ability to manage the permissions of an API user via the
API user edit page.