Add application js to srcFiles for jasmine tests
I want to introduce a new javascript module and some jasmine
tests. The current javascript modules under
`app/assets/javascripts/modules` are bundled into the
`application-<some-id>.js` asset when `assets:precompile` is
run. Precompiling assets is a pre-requisite for the `jasmine:ci`
command that runs our tests so this file needs to be added to the
`srcFiles` array so that it is picked up by jasmine's test runner.