Button to clear pre-selected permissions on /users/invitation/new
In #2436 we ensured that the default permissions for new users were
checked in the permissions checkboxes on /users/invitation/new. This
is a big time-saver for some inviting users who frequently add these
permissions. However there's a subset of inviting users who set up new
user accounts who don't need these default permissions (e.g. when
setting up users who use the Licensing app).

This commit adds a "Clear selected permissions" button above the list
of checkboxes which unchecks all the permissions checkboxes beneath
it.

I've decided to implement this using javascript. The
ClearSelectedPermissions class finds any `govuk-checkboxes__input`s
within the module and unchecks them when the button is clicked. The
javascript `preventDefault`s the buttons click event so that we don't
leave the page when it is clicked.

Since these checkboxes are contained within an `option-select`
component I've also taken advantage of the fact that this component
listens to a `change` event on elements with class
`gem-c-checkboxes__list`[1] to update the "selected count" above each
list of checkboxes.

In the test setup I've reproduced the minimum amount of HTML
(which in reality is generated by the govuk publishing components
gem's helpers) to verify the important behaviour of this module.

[1] https://github.com/alphagov/govuk_publishing_components/blob/8010b1321d3747f73d39e0616dfe0d781d8e43b1/app/assets/javascripts/govuk_publishing_components/components/option-select.js#L52