Do not display API-only apps on two more pages
Trello: https://trello.com/c/kvmb5OHO

This affects the following two pages:
* /users/:id/edit
* /account/manage_permissions

Managing editors should not need to see permissions for API-only apps
for non-API users or themselves and so we can reduce the noise on these
pages by excluding API-only apps as we already do for retired apps.

Note that we'll still be able to manage API user permissions on the API
users edit page.