Improve test for API user edit page
The main change is to ensure the name & email inputs are within the
right form since there's potentially more than one form on the page.

I've also extracted a setup block, because I'm planning to add more
tests in this context in subsequent commits.

I've also removed an unnecessary call to `ApiUser#id`; we can rely on
the test calling `ApiUser#to_param` and this makes the test more
idiomatic.