Make condition in API user edit page more consistent
The loop in the `app/views/shared/_user_permissions.html.erb` partial
calls `UserPermissionsControllerMethods#applications_and_permissions`
and so although calling `ApiUser#authorisations` probably gives the same
outcome in most/all cases, I think it's less confusing to call the same
method to decide whether to show that partial.

Also since we know the return value will be an Array, it feels more
readable and idiomatic to call `Enumerable#any?` rather than
`Object#present?`.

Note that I'm confident in making this change, because I added test
coverage in the "not allow editing permissions for application which
user does not have access to" test in `ApiUsersControllerTest` in an
earlier commit.