Add test for k8s sync_app_secrets task for retired app
The behaviour of this rake task was changed when we added `not_retired`
to the default scope on `Doorkeeper::Application` in this commit [1]
in #2446. Since then it has no longer synced app secrets for retired
apps; instead it will raise an exception.

This behaviour makes sense to me from a security point-of-view and I
have confirmed that (even after the recent spate of app retirements) the
Signon app still includes all the apps listed in
`charts/app-config/templates/signon-secrets-sync-configmap.yaml` [2]
which are used in the `sync-app-secrets-to-k8s` cron task e.g. in
integration [3]. This means that despite the change in behaviour we
should *not* see any exceptions. And if we do start seeing exceptions
that probably means the list of apps in
`signon-secrets-sync-configmap.yaml` needs to be updated.

[1]: 9a9cd02d8574529b19f572a8352f031c3f07455e
[2]: https://github.com/alphagov/govuk-helm-charts/blob/4c540acfa83899f0d73f9b06d480c712045cfc5f/charts/app-config/templates/signon-secrets-sync-configmap.yaml#L6-L42
[3]: https://github.com/alphagov/govuk-helm-charts/blob/4c540acfa83899f0d73f9b06d480c712045cfc5f/charts/app-config/values-integration.yaml#L2384-L2387