Add tests for SSO Push jobs for retired apps
This covers the scenario where an app is retired *after* the job is
enqueued but *before* it is executed.

The behaviour being tested was actually introduced when we added the
`not_retired` scope to the default scope for `Doorkeeper::Application`
in this commit [1] in #2446. This is because both jobs call
`Doorkeeper::Application.find_by` in their `#perform` methods which will
include the default scope and so if the app is retired will return `nil`
and be caught by the guard condition.

[1]: 9a9cd02d8574529b19f572a8352f031c3f07455e