Don't create/return SSO Push token for retired app
The logic in `PushUserUpdatesJob.perform_on`,
`PermissionUpdater#perform` & `ReauthEnforcer#perform` should mean that
`SSOPushCredential.credentials` (tested in previous commits) is never
called for a retired app.  However, on the off-chance that it ever is,
we never want to create a new access token or return one to be used by
`SSOPushClient#initialize`.