Remove del tags for retired apps on permissions pages
The following two tests give me confidence that retired apps no longer
appear in either of these tables and so the code to wrap the app name in
`del` tags is redundant.

* https://github.com/alphagov/signon/blob/2ddf47877fc3bde01b76dd476e92db1b9fc4b06b/test/controllers/users_controller_test.rb#L545-L559
* https://github.com/alphagov/signon/blob/2ddf47877fc3bde01b76dd476e92db1b9fc4b06b/test/controllers/account/manage_permissions_controller_test.rb#L34-L46

We also want to get rid of these tags in general, because there are a11y
problems with them.