Remove use of deprecated GovukStatsd
The use of this class is deprecated [1] in favour of Prometheus.

According to these docs [2], the data was stored in Graphite [3] but
that no longer seems to be accessible.

Since the data is no longer accessible, there's no point in trying to
send more data to Graphite. And these calls are cluttering up the code
for no good reason.

If we want to use Prometheus instead, we can do that as a separate step
and these calls to `GovukStatsd` will be in the git history if we need
to refer to them.

[1]: https://github.com/alphagov/govuk_app_config/blob/d83f88005b767765430c8a9a3e067ff69c40081c/README.md#statsd-deprecated
[2]: https://docs.publishing.service.gov.uk/manual/metrics.html#graphite
[3]: https://graphite.production.govuk.digital/