Remove unused export_usage_stats rake task
When I try to run this rake task in the integration environment, a
`TypeError` [1] is raised because `Rails.root` returns a `Pathname`
which cannot be converted into a String. `Rails.root` has returned a
`Pathname` since at least Rails v3 and so this hasn't worked for a long
time.

I can avoid the `TypeError` by explicitly setting the `GOVUK_APP_ROOT`
env var to "/app". However, then a `Errno::EROFS` exception [2] is
raised because it tries to write to the local filesystem and the EKS
platform containers have a read-only filesystem.

This means that the rake task hasn't worked for at least the best part
of a year since the Signon app was moved to the EKS platform. This in
turn suggests that the data in the generated CSV files were not that
important and so I think it's better to delete this unused code for now.

We already have an upcoming Trello card [3] about collecting some
metrics. I'll mention this code on that card in case there's some useful
stuff in these CSV files.

[1]: https://govuk.sentry.io/issues/4574177697/?project=202251
[2]: https://govuk.sentry.io/issues/4574185253/?project=202251
[3]: https://trello.com/c/2SRZa6cA