Remove unused script/permissions_by_organisation_csv_generator.rb
I haven't bothered trying to run this in the integration environment,
because I can see that it tries to write to a local CSV file which is no
longer possible on the EKS platform, because its containers have a
read-only filesystem.

This means that the rake task hasn't worked for at least the best part
of a year since the Signon app was moved to the EKS platform. This in
turn suggests that the data in the generated CSV files were not that
important and so I think it's better to delete this unused code for now.

We already have an upcoming Trello card [1] about collecting some
metrics. I'll mention this code on that card in case there's some useful
stuff in these CSV files.

[1]: https://trello.com/c/2SRZa6cA