Fix a test in EventLogPageIntegrationTest
This test became essentially useless in this commit [1] when we changed
the "Change your email or password" link on the home page to take you to
a new account page for changing email or password rather than the
admin-y user edit page.

The test became useless, because the assertion is testing a negative
and so while it's true there no link to the access log on the account
change email or password page, it never made sense to check for it on
that page.

I've made the test a bit more useful again by just trying to visit the
admin-y user edit page which a normal user shouldn't have access to.

As an aside it looks as if in this commit [2] we've possibly
inadvertently opened up a user's access log to users who couldn't
previously access it. However, I'm not going to address that here.

[1]: ad6d785b1577cfee227dced99b48abc6d271ce08
[2]: 5c35d13c3e720f60af4c89ea01ce876fa824388d