Log an event when a user successfully authenticates
We've had some requests from other teams who want to know how often
users log in to the applications they maintain.

We've decided to create an EventLog record each time an access token
is successfully authorized. This should be sufficient to answer the
question "how many unique users have accessed a specific application
in a given time period?".

Doorkeeper provides an `after_successful_authorization` hook which
we've used here to log an event after the authorization has completed
successfully. We considered adding this event logging in an
`after_create` hook on Doorkeeper::Token but decided against
re-opening that gem-provided class and relying instead on the hook
Doorkeeper provides.

We considered using the Doorkeeper::AccessToken records
themselves to derive the required metrics, but these tokens are issued
with a short expiry time and are periodically cleaned up by the
`oauth_access_records:delete_expired` task. It didn't feel appropriate
to overload these token records with reporting concerns.

We also considered storing this information at a less granular level
of detail (for example, only storing the last time a user accessed a
specific application rather than every time), but felt on balance that
because we have an existing EventLog mechanism, and the event log data
is cleaned up automatically by the
`event_log:delete_logs_older_than_two_years` task, that this approach
was the most consistent way of recording it.

Note that the gds-sso gem, which applications include in order to be
able to make this sequence of calls, uses Warden to cache the
logged-in status[1] for 20 hours[2], so we will only log these
requests per-user at most once every 20 hours.

[1] https://github.com/alphagov/gds-sso/blob/7f2a50dcc3bb73b7bb691742548f5a2423e7c5cc/lib/gds-sso/warden_config.rb#L31

[2] https://github.com/alphagov/gds-sso/blob/7f2a50dcc3bb73b7bb691742548f5a2423e7c5cc/lib/gds-sso/config.rb#L25

Co-authored-by: James Mead <james@floehopper.org>